+++ src/main/java/com/global/api/entities/enums/CountryCodeFormat.java
@@ -1,0 +1,8 @@
+package com.global.api.entities.enums;
+
+public enum CountryCodeFormat {
+    Alpha2,
+    Alpha3,
+    Numeric,
+    Name
+}